# backend/internal/httpapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Enterprise License Manager API",
    "version": "1.0.0",
    "description": "API quản lý người dùng, thiết bị, phần mềm, license, cấp phát, audit log, dashboard và self-service. Các endpoint quản trị yêu cầu JWT và kiểm tra vai trò ở backend."
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/licenses": {
      "get": {
        "tags": ["Licenses"],
        "operationId": "listLicenses",
        "summary": "Danh sách license và số seat",
        "x-required-roles": ["admin", "it_manager"],
        "responses": { "200": { "description": "Danh sách license", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseList" } } } }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["Licenses"],
        "operationId": "createLicense",
        "summary": "Tạo license",
        "x-required-roles": ["admin", "it_manager"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseInput" } } } },
        "responses": { "201": { "description": "License đã tạo; response không chứa plaintext key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/License" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/licenses/{id}": {
      "put": {
        "tags": ["Licenses"],
        "operationId": "updateLicense",
        "summary": "Cập nhật license",
        "description": "Gửi license_key rỗng để giữ nguyên key hiện tại.",
        "x-required-roles": ["admin", "it_manager"],
        "parameters": [{ "$ref": "#/components/parameters/EntityID" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseInput" } } } },
        "responses": { "200": { "description": "License đã cập nhật", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/License" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
...
    "/api/v1/me/licenses": {
      "get": {
        "tags": ["Self-service"],
        "operationId": "getMyLicenses",
        "summary": "License của tài khoản hiện tại",
        "description": "Bao gồm license cấp trực tiếp và qua thiết bị hiện đang được giao. can_view_key cho biết Employee có thể yêu cầu giải mã key hay không.",
        "responses": { "200": { "description": "License của tôi", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssignedLicenseList" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
...
    "/api/v1/me/requestable-software": {
      "get": {
        "tags": ["License requests"],
        "operationId": "getRequestableSoftware",
        "summary": "Danh mục phần mềm Employee có thể yêu cầu",
        "x-required-roles": ["employee"],
        "responses": { "200": { "description": "Danh mục phần mềm", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SoftwareList" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/me/license-requests": {
      "get": {
        "tags": ["License requests"],
        "operationId": "getMyLicenseRequests",
        "summary": "Lịch sử yêu cầu license của Employee hiện tại",
        "x-required-roles": ["employee"],
        "responses": { "200": { "description": "Yêu cầu của tôi", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequestList" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["License requests"],
        "operationId": "createMyLicenseRequest",
        "summary": "Tạo yêu cầu cấp license cho bản thân",
        "x-required-roles": ["employee"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequestCreate" } } } },
        "responses": { "201": { "description": "Yêu cầu đã tạo", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequest" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/me/license-requests/{id}/cancel": {
      "patch": {
        "tags": ["License requests"],
        "operationId": "cancelMyLicenseRequest",
        "summary": "Hủy yêu cầu pending của bản thân",
        "x-required-roles": ["employee"],
        "parameters": [{ "$ref": "#/components/parameters/EntityID" }],
        "responses": { "200": { "description": "Yêu cầu đã hủy", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequest" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/license-requests": {
      "get": {
        "tags": ["License requests"],
        "operationId": "listLicenseRequests",
        "summary": "Danh sách yêu cầu cho Admin và IT Manager",
        "x-required-roles": ["admin", "it_manager"],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "approved", "rejected", "cancelled"] } },
          { "name": "priority", "in": "query", "schema": { "type": "string", "enum": ["normal", "high", "urgent"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Danh sách yêu cầu", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequestList" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/license-requests/{id}/approve": {
      "patch": {
        "tags": ["License requests"],
        "operationId": "approveLicenseRequest",
        "summary": "Duyệt yêu cầu và tự động cấp license",
        "x-required-roles": ["admin", "it_manager"],
        "parameters": [{ "$ref": "#/components/parameters/EntityID" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequestApprove" } } } },
        "responses": { "200": { "description": "Yêu cầu đã duyệt", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequest" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "422": { "$ref": "#/components/responses/UnprocessableEntity" } }
      }
    },
    "/api/v1/license-requests/{id}/reject": {
      "patch": {
        "tags": ["License requests"],
        "operationId": "rejectLicenseRequest",
        "summary": "Từ chối yêu cầu và gửi phản hồi",
        "x-required-roles": ["admin", "it_manager"],
        "parameters": [{ "$ref": "#/components/parameters/EntityID" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequestReject" } } } },
        "responses": { "200": { "description": "Yêu cầu đã từ chối", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseRequest" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "EntityID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
    },
    "responses": {
      "BadRequest": { "description": "Dữ liệu request không hợp lệ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Thiếu token hoặc token không hợp lệ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Không đủ quyền", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Không tìm thấy dữ liệu", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "Xung đột quy tắc nghiệp vụ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "UnprocessableEntity": { "description": "Đối tượng tham chiếu hoặc trạng thái không phù hợp", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } } },
...
      "SoftwareProduct": {
        "type": "object",
        "required": ["id", "name", "publisher", "version", "description", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" }, "publisher": { "type": "string" },
          "version": { "type": "string" }, "description": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "License": {
        "type": "object",
        "required": ["id", "software_product_id", "name", "license_type", "assignment_type", "seat_count", "used_seats", "available_seats", "allow_employee_key_view", "cost", "lifecycle_status", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" }, "software_product_id": { "type": "string", "format": "uuid" }, "name": { "type": "string" },
          "license_type": { "type": "string", "enum": ["subscription", "perpetual"] },
          "assignment_type": { "type": "string", "enum": ["user", "device", "mixed"] },
          "seat_count": { "type": "integer", "minimum": 1 }, "used_seats": { "type": "integer" }, "available_seats": { "type": "integer" },
          "key_hint": { "type": "string", "readOnly": true }, "allow_employee_key_view": { "type": "boolean", "default": false }, "vendor": { "type": "string" },
          "purchased_at": { "type": "string", "format": "date" }, "starts_at": { "type": "string", "format": "date" }, "expires_at": { "type": "string", "format": "date" },
          "cost": { "type": "number", "format": "double" }, "currency": { "type": "string", "minLength": 3, "maxLength": 3 }, "notes": { "type": "string" },
          "lifecycle_status": { "type": "string", "enum": ["upcoming", "active", "expired", "archived"] },
          "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" },
          "archived_at": { "type": "string", "format": "date-time", "nullable": true, "readOnly": true }
        }
      },
      "LicenseInput": {
        "type": "object",
        "required": ["software_product_id", "name", "license_type", "assignment_type", "seat_count"],
        "properties": {
          "software_product_id": { "type": "string", "format": "uuid" }, "name": { "type": "string" },
          "license_type": { "type": "string", "enum": ["subscription", "perpetual"] }, "assignment_type": { "type": "string", "enum": ["user", "device", "mixed"] },
          "seat_count": { "type": "integer", "minimum": 1 }, "license_key": { "type": "string", "writeOnly": true }, "allow_employee_key_view": { "type": "boolean", "default": false }, "vendor": { "type": "string" },
          "purchased_at": { "type": "string", "format": "date" }, "starts_at": { "type": "string", "format": "date" }, "expires_at": { "type": "string", "format": "date" },
          "cost": { "type": "number", "minimum": 0 }, "currency": { "type": "string", "minLength": 3, "maxLength": 3 }, "notes": { "type": "string" }
        }
      },
...
      "AssignedLicense": {
        "type": "object",
        "properties": {
          "assignment_id": { "type": "string", "format": "uuid" }, "license_id": { "type": "string", "format": "uuid" }, "license_name": { "type": "string" },
          "software_product_id": { "type": "string", "format": "uuid" }, "license_type": { "type": "string" }, "assignment_source": { "type": "string", "enum": ["user", "device"] },
          "device_id": { "type": "string", "format": "uuid" }, "device_asset_code": { "type": "string" }, "assigned_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date" }, "lifecycle_status": { "type": "string", "enum": ["upcoming", "active", "expired", "archived"] }, "notes": { "type": "string" }, "can_view_key": { "type": "boolean", "readOnly": true }
        }
      },
      "LicenseRequest": {
        "type": "object",
        "required": ["id", "requester_id", "requester_name", "software_product_id", "software_product_name", "priority", "reason", "status", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" }, "requester_id": { "type": "string", "format": "uuid" }, "requester_name": { "type": "string" },
          "software_product_id": { "type": "string", "format": "uuid" }, "software_product_name": { "type": "string" },
          "priority": { "type": "string", "enum": ["normal", "high", "urgent"] }, "reason": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "approved", "rejected", "cancelled"] },
          "selected_license_id": { "type": "string", "format": "uuid" }, "selected_license_name": { "type": "string" }, "assignment_id": { "type": "string", "format": "uuid" },
          "reviewed_by": { "type": "string", "format": "uuid" }, "reviewed_by_name": { "type": "string" },
          "decision_reason": { "type": "string", "enum": ["out_of_stock", "not_approved", "other"] }, "response_note": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" },
          "reviewed_at": { "type": "string", "format": "date-time" }, "cancelled_at": { "type": "string", "format": "date-time" }
        }
      },
      "LicenseRequestCreate": { "type": "object", "required": ["software_product_id", "priority", "reason"], "properties": { "software_product_id": { "type": "string", "format": "uuid" }, "priority": { "type": "string", "enum": ["normal", "high", "urgent"] }, "reason": { "type": "string" } } },
      "LicenseRequestApprove": { "type": "object", "required": ["license_id"], "properties": { "license_id": { "type": "string", "format": "uuid" }, "response_note": { "type": "string" } } },
      "LicenseRequestReject": { "type": "object", "required": ["decision_reason", "response_note"], "properties": { "decision_reason": { "type": "string", "enum": ["out_of_stock", "not_approved", "other"] }, "response_note": { "type": "string" } } },
...
      "SoftwareList": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/SoftwareProduct" } } } }] },
      "LicenseList": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/License" } } } }] },
...
      "AssignedLicenseList": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/AssignedLicense" } } } }] },
      "LicenseRequestList": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/LicenseRequest" } } } }] },
...
      "ListEnvelope": { "type": "object", "required": ["items", "total"], "properties": { "items": { "type": "array", "items": {} }, "total": { "type": "integer", "minimum": 0 } } }
    }
  }
}